make it so that both the creation/update pages use new stuff
